pyproject.toml
[project]
name = "image-caption-playground"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow>=12.0.0",
    "streamlit>=1.51.0",
    "torch>=2.9.0",
    "transformers>=4.57.1",
]
